[frontend] extract gate graph in its own module (#200)
Also shuffle things around to their natural places.

One thing that we had to do due to the privacy issues is to use SecondaryMap
instead of Vec<ValueIndex> for tracking the correspondence between Wire and
ValueIndex.
