+++ crates/frontend/src/compiler/gate/assert_band_0.rs
@@ -10,9 +10,11 @@
 ///
 /// The gate generates 1 AND constraint:
 /// - `x ∧ constant = 0`
-use super::{Gate, GateData};
 use crate::{
-	compiler::circuit,
+	compiler::{
+		circuit,
+		gate_graph::{Gate, GateData},
+	},
 	constraint_system::{AndConstraint, ConstraintSystem},
 	word::Word,
 };
