+++ crates/frontend/src/compiler/gate/assert_eq_cond.rs
@@ -12,9 +12,11 @@
 ///
 /// The gate generates 1 AND constraint:
 /// - `(x ⊕ y) ∧ mask = 0`
-use super::{Gate, GateData};
 use crate::{
-	compiler::circuit,
+	compiler::{
+		circuit,
+		gate_graph::{Gate, GateData},
+	},
 	constraint_system::{AndConstraint, ConstraintSystem},
 	word::Word,
 };
