+++ crates/frontend/src/compiler/gate/icmp_eq.rs
@@ -21,9 +21,11 @@
 /// The gate generates two AND constraints:
 /// 1. Carry propagation: `(x ⊕ y ⊕ cin) ∧ (all-1 ⊕ cin) = cin ⊕ cout`
 /// 2. Mask generation: `out_mask = (cout SRA 63) ⊕ all-1`
-use super::{Gate, GateData};
 use crate::{
-	compiler::circuit,
+	compiler::{
+		circuit,
+		gate_graph::{Gate, GateData},
+	},
 	constraint_system::{AndConstraint, ConstraintSystem, ShiftedValueIndex},
 	word::Word,
 };
