+++ crates/frontend/src/compiler/gate/icmp_ult.rs
@@ -19,9 +19,11 @@
 /// The gate generates 2 AND constraints:
 /// 1. Borrow propagation: `(¬x ⊕ bin) ∧ (y ⊕ bin) = bin ⊕ bout`
 /// 2. Mask generation: `out_mask = bout SRA 63`
-use super::{Gate, GateData};
 use crate::{
-	compiler::circuit,
+	compiler::{
+		circuit,
+		gate_graph::{Gate, GateData},
+	},
 	constraint_system::{AndConstraint, ConstraintSystem, ShiftedValueIndex},
 	word::Word,
 };
