+++ crates/frontend/src/compiler/gate/shr.rs
@@ -11,9 +11,11 @@
 ///
 /// The gate generates 1 AND constraint:
 /// - `(x >> n) ∧ all-1 = z`
-use super::{Gate, GateData};
 use crate::{
-	compiler::circuit,
+	compiler::{
+		circuit,
+		gate_graph::{Gate, GateData},
+	},
 	constraint_system::{AndConstraint, ConstraintSystem, ShiftedValueIndex},
 };
