+++ crates/frontend/src/compiler/gate/shr32.rs
@@ -10,9 +10,11 @@
 ///
 /// The gate generates 1 AND constraint:
 /// - `(x >> n) ∧ MASK_32 = z`
-use super::{Gate, GateData};
 use crate::{
-	compiler::circuit,
+	compiler::{
+		circuit,
+		gate_graph::{Gate, GateData},
+	},
 	constraint_system::{AndConstraint, ConstraintSystem, ShiftedValueIndex},
 };
